refactor(AddRewardButton): extract constants & types to separate files
